TODO feat: export openapi spec generated with aide and schemars
Export an openapi spec for the VTN server with the aide and schemars crates.

---

This is a draft commit that explores the aide and schemars crates.

The aide crate is not as popular as utoipa. It specifically targets the
axum web framework and relies on the (popular) schemars crate for
generating json schemas that honors serialization (and validation)
defined with the serde (and validator) crate.

Together, aide and schemars provide macros, traits, and types for
generating openapi documentation.

This draft generates documentation for two operations

- GET /vens/{venID}/resources
- POST /vens/{venID}/resources

The document was exported to generated-openapi.yaml with

    curl --silent http://localhost:3000/docs/openapi.json | yq --yaml-output . > generated-openapi.yaml

Some callouts

- #[serde(flatten)] is handled as expected
- the effect of other serde attributes and validator constraints are
  handled as expected
- the crate sometimes encourages or forces you to slightly alter the
  shape of your types in order to work with it -- for example, path
  parameters must be a named field in a struct in order to automatically
  infer the path parameter's name in documentation
- aide uses openapi verson 3.1.0

Overall, using the aide and schemars crates is a nice experience. It
does require boilerplate, but that boilerplate is largely vanilla rust
code, which is nicer to write and maintain than macro usage in my
opinion.

See https://docs.rs/aide/latest/aide/
and https://crates.io/crates/aide
and https://docs.rs/schemars/latest/schemars/
and https://crates.io/crates/schemars
and https://www.openapis.org/blog/2021/02/16/migrating-from-openapi-3-0-to-3-1-0