Fix Dockerfile to support MSRV 1.81 (#72)
* Fix Dockerfile to support MSRV 1.81

Fix the Dockerfile such that it's based on a rust image that meets the
minimum supported rust version (MSRV) defined in cargo (1.81).

Signed-off-by: Ben Webb <benjaminedwardwebb@gmail.com>

* Add CI step that builds Docker image

Add CI step that checks that the vtn.Dockerfile successfully builds a
docker image on each committed change.

We opt to not publish the image at this point in time.

See the prior commit for some motivation.

Refs: 5c188e54eb507998691d377c564e1512eab6f00f
Signed-off-by: Ben Webb <benjaminedwardwebb@gmail.com>

* Separate docker build from CI build job

Signed-off-by: Maximilian Pohl <maximilian@tweedegolf.com>

---------

Signed-off-by: Ben Webb <benjaminedwardwebb@gmail.com>
Signed-off-by: Maximilian Pohl <maximilian@tweedegolf.com>
Co-authored-by: Maximilian Pohl <maximilian@tweedegolf.com>